+++ kernel/cutlass/test/unit/gemm/device/testbed_rank_k_universal.h
@@ -1,5 +1,5 @@
 /***************************************************************************************************
- * Copyright (c) 2017 - 2023 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+ * Copyright (c) 2017 - 2024 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
  * SPDX-License-Identifier: BSD-3-Clause
  *
  * Redistribution and use in source and binary forms, with or without
@@ -272,7 +272,7 @@ struct TestbedRank2KUniversal {
     // Determine SMEM requirements and waive if not satisfied
     //
 
-    int smem_size = int(sizeof(typename RankK::RankKkernel::SharedStorage));
+    size_t smem_size = sizeof(typename RankK::RankKkernel::SharedStorage);
 
     cudaDeviceProp properties;
     int device_idx;
