+++ kernel/cutlass/test/unit/gemm/device/testbed_symm_universal.h
@@ -1,5 +1,5 @@
 /***************************************************************************************************
- * Copyright (c) 2017 - 2023 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+ * Copyright (c) 2017 - 2024 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
  * SPDX-License-Identifier: BSD-3-Clause
  *
  * Redistribution and use in source and binary forms, with or without
@@ -311,7 +311,7 @@ struct TestbedSymmUniversal {
     // Determine SMEM requirements and waive if not satisfied
     //
 
-    int smem_size = int(sizeof(typename Symm::SymmKernel::SharedStorage));
+    size_t smem_size = sizeof(typename Symm::SymmKernel::SharedStorage);
 
     cudaDeviceProp properties;
     int device_idx;
